requirements.txt
yt-dlp>=2025.1.0
openai>=1.65.0
python-dotenv>=1.0.1
